requirements.txt
streamlit
openai
tweepy
textblob
chromadb
python-dotenv
pandas
